Update solution for majority-classifier

+++ majority-classifier/majority-classifier.py
@@ -10,5 +10,5 @@ def majority_classifier(y_train, X_test):
     majority_label = labels[np.argmax(counts)]
     if len(X_test) == 0:
        return np.asarray([], dtype = int)
-    predictions = np.full(shape = (len(X_test), ), fill_value = majority_label, dtype = int)
+    predictions = np.full(shape = (len(X_test), ), fill_value = majority_label)
     return predictions
